# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Fazaa membership registration
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegistrationInputBrand = typeof RegistrationInputBrand[keyof typeof RegistrationInputBrand];


export const RegistrationInputBrand = {
  fazaa: 'fazaa',
  esaad: 'esaad',
  homat: 'homat',
  alsaada: 'alsaada',
  absher: 'absher',
} as const;

export type RegistrationInputCardType = typeof RegistrationInputCardType[keyof typeof RegistrationInputCardType];


export const RegistrationInputCardType = {
  gold: 'gold',
  silver: 'silver',
  discount: 'discount',
} as const;

export type RegistrationInputPaymentMethod = typeof RegistrationInputPaymentMethod[keyof typeof RegistrationInputPaymentMethod];


export const RegistrationInputPaymentMethod = {
  card: 'card',
  apple_pay: 'apple_pay',
} as const;

export interface RegistrationInput {
  fullName: string;
  phone: string;
  emiratesId: string;
  brand: RegistrationInputBrand;
  cardType: RegistrationInputCardType;
  region: string;
  streetAddress: string;
  neighborhood: string;
  deliveryDate: string;
  paymentMethod: RegistrationInputPaymentMethod;
}

export type StatusUpdateStatus = typeof StatusUpdateStatus[keyof typeof StatusUpdateStatus];


export const StatusUpdateStatus = {
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface StatusUpdate {
  status: StatusUpdateStatus;
}

export interface Registration {
  id: number;
  fullName: string;
  phone: string;
  emiratesId: string;
  brand: string;
  cardType: string;
  region: string;
  streetAddress: string;
  neighborhood: string;
  deliveryDate: string;
  paymentMethod: string;
  status: string;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}