# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification for Fazaa membership registration
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: registrations
    description: Membership registration operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /registrations:
    post:
      operationId: createRegistration
      tags: [registrations]
      summary: Submit a membership registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listRegistrations
      tags: [registrations]
      summary: List all registrations (admin)
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Registration"
  /registrations/{id}:
    get:
      operationId: getRegistration
      tags: [registrations]
      summary: Get registration by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Registration found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /registrations/{id}/status:
    patch:
      operationId: updateRegistrationStatus
      tags: [registrations]
      summary: Approve or reject a registration (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StatusUpdate"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    RegistrationInput:
      type: object
      required:
        - fullName
        - phone
        - emiratesId
        - brand
        - cardType
        - region
        - streetAddress
        - neighborhood
        - deliveryDate
        - paymentMethod
      properties:
        fullName:
          type: string
        phone:
          type: string
        emiratesId:
          type: string
        brand:
          type: string
          enum: [fazaa, esaad, homat, alsaada, absher]
        cardType:
          type: string
          enum: [gold, silver, discount]
        region:
          type: string
        streetAddress:
          type: string
        neighborhood:
          type: string
        deliveryDate:
          type: string
        paymentMethod:
          type: string
          enum: [card, apple_pay]
    StatusUpdate:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [approved, rejected]
    Registration:
      type: object
      required:
        - id
        - fullName
        - phone
        - emiratesId
        - brand
        - cardType
        - region
        - streetAddress
        - neighborhood
        - deliveryDate
        - paymentMethod
        - status
        - createdAt
      properties:
        id:
          type: integer
        fullName:
          type: string
        phone:
          type: string
        emiratesId:
          type: string
        brand:
          type: string
        cardType:
          type: string
        region:
          type: string
        streetAddress:
          type: string
        neighborhood:
          type: string
        deliveryDate:
          type: string
        paymentMethod:
          type: string
        status:
          type: string
        createdAt:
          type: string
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string